requirements.txt
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
cffi==1.11.2
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.1.1
parsel==1.2.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.5.1
pyOpenSSL==17.3.0
pypiwin32==220
python-slugify==1.2.4
queuelib==1.4.2
Scrapy==1.4.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
Unidecode==0.4.21
w3lib==1.18.0
zope.interface==4.4.3
